feat: getUserProfile 함수 추가
